Adding "dynamic mode", splitting out enabled/disable logic into it's own file
This adds the ability for namespaces to be enabled or disabled on the fly (aka: dynamic mode)
Rather than simply returning an enabled/disabled function, when dynamic mode is engaged, it
will instead return a thin wrapper that checks the enabled status before each and every call.

To accomodate this, the enable/disable lists needed to be upgraded. The new file (able.js)
is a singleton that facilitates this feature. Before, the list was really only capable of being
modified during init, but afterwards fell flat. This brings more capability there, and also
makes that specific feature more testable.