Make millisecond timer namespace specific
When debugging node apps, I find it much more useful for the millisecond
timer to be relative to last message from the same namespace instead of
any message. This is especially true when I'm debugging across multiple
libraries or multiple levels in the same module and I'm interested in seeing
all the messages but also need to compare times from specific levels.