Expose inspect to allow custom object formatting
This commit exposes the inspect method, which is used to format objects
for the `%o` formatting directive, so that it can be called or replaced
to suit the needs of users.

Some use cases for this are #185 (inspect with same coloring as debug)
and #263 ("better" object formatting).  My primary use case is to log
objects with a smaller depth to avoid spamming the logs from large/deep
objects.  Other use cases include using the showHidden and customInspect
options of util.inspect.

Why not just use `util.inspect` or another custom formatting function
directly?  My major reasons are:

* It takes more code since every debugged object must be formatted.
* It's not available on browsers, complicating Universal/Isomorphic
  JavaScript (or interfering with the inspector if polyfilled).
* It adds unnecessary formatting overhead when debugging is disabled.
* It doesn't match debug output colorization without using the
  undocumented .useColors property.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>