Fix devtools console formatting
babel explicitly requires `debug/node.js`
via hadron-compile-cache in the renderer
which then causes all messages in the devtools
console to show up in an unreadable form.

So, just like in `index.js`, we also just check
in `node.js` as well if we're actually in a renderer
process and swap module.exports.