Fix implicit window access
The partial fix before did not address accessing `navigator`.  I whipped this up in the Github editor so you might want to reformat it.  Totally appreciate the work you guys do.  That said, this component is used by a lot of modules, in a lot of situations.  This is the 2nd week in a row where debug has broken my build process.  Expanding the test suite to test with browser and non-browser situations would be great.